+++ README.md
@@ -1,5 +1,5 @@
 # TapResearch-iOS-SDK
-TapResearch iOS SDK v2.3.0
+TapResearch iOS SDK v2.3.0u
 
 
 For additional information, please see the [TapResearch iOS SDK integration guide](https://www.tapresearch.com/docs/ios_integration_guide).
@@ -13,7 +13,7 @@ Create an [app](/supplier_dashboard/dashboard/apps/new) and grab your API Token.
 
 Add the pod info to the app's Podfile
  ```ruby
- pod 'TapResearch', '2.3.0'
+ pod 'TapResearch', '2.3.0u'
  ```
 
  Then run the following command:
