+++ TapResearch.podspec
@@ -1,6 +1,6 @@
 Pod::Spec.new do |s|
   s.name             = 'TapResearch'
-  s.version          = '2.3.0'
+  s.version          = '2.3.0u'
   s.summary          = 'TapResearch IOS SDK'
 
   s.description      = <<-DESC
@@ -10,9 +10,9 @@ TapResearch IOS SDK, to learn more checkout: https://www.tapresearch.com/docs/io
   s.homepage         = 'https://www.tapresearch.com'
   s.license          = { :type => 'Closed Source' }
   s.author           = { 'Jeroen Verbeek' => 'jeroen@tapresarch.com' }
-  s.source           = { :git => 'https://github.com/TapResearch/TapResearch-iOS-SDK.git', :tag => 'v2.3.0' }
+  s.source           = { :git => 'https://github.com/TapResearch/TapResearch-iOS-SDK.git', :tag => 'v2.3.0u' }
 
   s.ios.deployment_target = '9.2'
-  s.ios.vendored_frameworks = 'TapResearchSDK.xcframework'
+  s.ios.vendored_frameworks = 'TapResearchSDK.framework'
   s.frameworks = 'UIKit', 'MapKit', 'Foundation', 'SystemConfiguration', 'MobileCoreServices', 'AdSupport', 'Security'
 end
