+++ TapResearchSDK.framework/TapResearchSDK
@@ -1,0 +1,1 @@
+Versions/Current/TapResearchSDK
