+++ TapResearchSDK.framework/Versions/A/Headers/TRPlacementCustomParameter.h
@@ -1,0 +1,20 @@
+//
+//  TRPlacementParameter.h
+//  TapResearchSDK
+//
+//  Created by Ilan Caspi on 1/16/20.
+//  Copyright © 2020 TapResearch. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+//NS_ASSUME_NONNULL_BEGIN
+
+@interface TRPlacementCustomParameter : NSObject
+
+@property (nonatomic, strong) NSString * _Nonnull key;
+@property (nonatomic, strong) NSString * _Nonnull value;
+
+@end
+
+//NS_ASSUME_NONNULL_END
