+++ TapResearchSDK.framework/Versions/A/Headers/TRPlacementCustomParameterList.h
@@ -1,0 +1,21 @@
+//
+//  TRPlacementCustomParameterList.h
+//  TapResearchSDK
+//
+//  Created by Ilan Caspi on 1/17/20.
+//  Copyright © 2020 TapResearch. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+#import "TRPlacementCustomParameter.h"
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface TRPlacementCustomParameterList : NSObject
+
+- (BOOL)addParameter:(TRPlacementCustomParameter *)customParameter;
+- (NSMutableDictionary *)toDictionary;
+
+@end
+
+NS_ASSUME_NONNULL_END
