+++ TapResearchSDK.framework/Versions/A/Headers/TRSerializationHelper.h
@@ -1,0 +1,17 @@
+//
+//  TRSerializationHelper.h
+//  TapResearchSDK
+//
+//  Created by Ilan Caspi on 3/5/18.
+//  Copyright © 2018 TapResearch. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+@interface TRSerializationHelper : NSObject
+
++ (NSDictionary *) dictionaryWithPropertiesOfObject:(id)obj;
++ (NSString *)jsonStringFromDictionary:(NSDictionary *)dictionary;
++ (NSString *)jsonStringFromArray:(NSArray *)array;
+
+@end
