+++ TapResearchSDK.framework/Versions/A/Headers/TapResearchSDK.h
@@ -1,0 +1,17 @@
+//
+//  TapResearchSDK.h
+//  TapResearchSDK
+//
+//  Copyright (c) 2015 TapResearch. All rights reserved.
+//
+
+#import "TapResearch.h"
+#import "TRPlacement.h"
+#import "TRReward.h"
+#import "TRSerializationHelper.h"
+#import "TRPlacementCustomParameter.h"
+#import "TRPlacementCustomParameterList.h"
+#import "TRPlacementCustomParameterBuilder.h"
+#import "TRPlacementCustomParameter+Builder.h"
+
+
