--- TapResearchSDK.xcframework/ios-arm64_armv7/TapResearchSDK.framework/Headers/TapResearchSDK.h
@@ -1,17 +1,0 @@
-//
-//  TapResearchSDK.h
-//  TapResearchSDK
-//
-//  Copyright (c) 2015 TapResearch. All rights reserved.
-//
-
-#import <TapResearchSDK/TapResearch.h>
-#import <TapResearchSDK/TRPlacement.h>
-#import <TapResearchSDK/TRReward.h>
-#import <TapResearchSDK/TRSerializationHelper.h>
-#import <TapResearchSDK/TRPlacementCustomParameter.h>
-#import <TapResearchSDK/TRPlacementCustomParameterList.h>
-#import <TapResearchSDK/TRPlacementCustomParameterBuilder.h>
-#import <TapResearchSDK/TRPlacementCustomParameter+Builder.h>
-
-
