--- TapResearchSDK.xcframework/ios-arm64_armv7/TapResearchSDK.framework/Modules/module.modulemap
@@ -1,6 +1,0 @@
-framework module TapResearchSDK {
-  umbrella header "TapResearchSDK.h"
-
-  export *
-  module * { export * }
-}
