--- TapResearchSDK.xcframework/ios-i386_x86_64-simulator/TapResearchSDK.framework/Headers/TRPlacementCustomParameter+Builder.h
@@ -1,17 +1,0 @@
-//
-//  TRPlacementCutsomParameter+Builder.h
-//  TapResearchSDK
-//
-//  Created by Ilan Caspi on 1/16/20.
-//  Copyright © 2020 TapResearch. All rights reserved.
-//
-
-#import <TapResearchSDK/TRPlacementCustomParameterBuilder.h>
-
-NS_ASSUME_NONNULL_BEGIN
-
-@interface TRPlacementCustomParameter(Builder)
--(TRPlacementCustomParameterBuilder*) builder;
-@end
-
-NS_ASSUME_NONNULL_END
