--- TapResearchSDK.xcframework/ios-i386_x86_64-simulator/TapResearchSDK.framework/Headers/TRPlacementCustomParameterBuilder.h
@@ -1,22 +1,0 @@
-//
-//  TRPlacementCustomParameterBuilder.h
-//  TapResearchSDK
-//
-//  Created by Ilan Caspi on 1/16/20.
-//  Copyright © 2020 TapResearch. All rights reserved.
-//
-
-#import <Foundation/Foundation.h>
-#import <TapResearchSDK/TRPlacementCustomParameter.h>
-
-NS_ASSUME_NONNULL_BEGIN
-
-@interface TRPlacementCustomParameterBuilder : NSObject
-- (instancetype)initWithPlacementCustomParameter:(TRPlacementCustomParameter *)parameter;
-- (instancetype)key:(NSString *)key;
-- (instancetype)value:(NSString *)value;
-- (TRPlacementCustomParameter*)build;
-
-@end
-
-NS_ASSUME_NONNULL_END
