--- TapResearchSDK.xcframework/ios-i386_x86_64-simulator/TapResearchSDK.framework/Modules/module.modulemap
@@ -1,6 +1,0 @@
-framework module TapResearchSDK {
-  umbrella header "TapResearchSDK.h"
-
-  export *
-  module * { export * }
-}
